change query nested conds to use #any? instead of #all?
It feels slightly more natural to me for this case to be asking for
posts with *any* popular tags rather than *all*

it doesn't break any tests which means we don't really specify the behavior yet.

some test cases around this could probably clarify the difference and help
us figure out if this is really the right way to go.